media-uploader/requirements.txt
sqlalchemy>=1.4.0,<2.0.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
python-dotenv>=0.19.0
PyJWT>=2.6.0
httpx>=0.24.0
python-jose>=3.3.0
cryptography>=40.0.0
fastapi-utils>=0.2.1
fastapi
uvicorn
pytest
minio
redis>=4.0.0
passlib[bcrypt]>=1.7.4
psutil>=5.9.0
